Fix bugs in the memory handle reduction logic of the uGNI transport.
The idea is to reuse memory handles by defining a min and max bound on
registered memory. When we register new memory, we check if the requested
address and size fits in the current region and if not expand the region to include
the new address. We need to save the old regions because they may have already
been advertised to a peer and are being used for RDMA operations.

Signed-off-by: Tom Tucker <tom@ogc.us>