Re-structured configure.ac and Makefile.am's.
This change contains the change in configure.ac and a set of Makefile.am's for
each module. Some source files are also modified to use 'coll/idx.h' instead of
'sos/idx.h'.

Additional M4 macros for configuration option handling is also added.

The code compiles, but not fully tested.