ldms: multi-threading store flush and configure.ac bug fix.
This changeset contains configure.ac bug fix and the implementation to support
multi-threading store flush. The modified program has been tested on vic1 and
vic2 and seems to work. (NOTE: The tester use gdb to ensure that the
ldmsd_store_flush is called and he is not delusional.)

ldmsd also accept another parameter '-f' for the number of flush threads. In
addition, ldmsctl command "info" now reports the flush threads' information
(number of store and time spent in flushing for each thread).

Change-Id: I801d9851c28a51f14d0c9e21fddb7ae65dc49df5