ldms: ugni_xprt fix
This fix takes care of the erroneous closing of fd(0) and assignment
of this descriptor to the store with subsequent closes. This also fixes
the problem, due to the same condition, of a sampler not being included in
an aggregators data set if the aggregator was started before the sampler.
Note that the parent of this change is the dlopen fix.

Change-Id: Ib8b088b7410e727b9aee04f0fe6fe1e32d32aa32