Destroy a set when the update fails to force a lookup.
When the update fails it implies the peer has destroyed that set; for example if
the peer is an aggregator and the compute node has gone away, the connection
will persist, but its sets will ultimately be destroyed (when it comes back online).
This change will allows LDMSD to recover and resume updating/storing data for
the set when it returns.

Signed-off-by: Tom Tucker <tom@ogc.us>