Fix sockets transport to properly detect set removal on the peer.
Prior to this change, the sockets transport send addresses back and forth over
the wire for "remote buffers." This is a problem because effectively this buffer
is never verified at the peer. If the set goes away, for example, the update will
still succeed because the address is still valid in the memory of the peer process.
Also, if the address is "bad", it will crash the peer when it attempts to access it.
This code adds logical memory keys with a lookup to validate the incoming request.
These keys are destroyed when the associated set is destroyed.

Signed-off-by: Tom Tucker <tom@ogc.us>