bulk: Bulk change.
This change contains the folowwing:

1. Refactor the code to use a single RBT and remove the other one
(ogc_rbt_xxx). These changes have been unit tested and the bpt.c test has
been made more robust.

2. Moved the dropping of a hset reference to the end of the function. This
was a bug, but I don't think it was symptomatic because we never remove hsets

3. Converted signal to sigaction in ldmsd.c. Should also specifically
print out the sig_addr. This would tell us the address that caused a
SIGFAULT for example even when not attached with a debugger.

4. Whitespace abuse annoyance

5. converted select to epoll for the connect timeout in the sock and ugni
transports.

Change-Id: I9f5385f65578a4d41cf1480dbc82a27154f12167