sos: key comparator fix
comparator(a,b) was defined as (a - b). This was OK if a and b are not much
different. In the case that a is much greater than b (e.g. a = MAX_UINT64 and b
is less than MAX_UINT64/2), a - b will overflow MAX_INT64, resulting in wrong
calculation. This changeset modify the comparator to do actual comparison and
returns -1, 0, 1 in the case of a is less than, equal to, and greater than b
respectively.

NOTE: The bug causes komondor to crash as komondor use sos_iter_seek_inf with
MAX_UINT64. This changeset fixes it.

Change-Id: I5aaf35d4eec36b80b395d4670679e4b2287a026e