ldms: sysclassib using MAD message to query perofrmance counter.
New implementation of sysclassib, using MAD messages to perform performance
counter and counter reset. The sampler also check whether each of the given
ports supports extended performance counters. If so, it will not reset the
counters, but query the extended counters in addition to regular counters. In
the case that the port does not support extended performance counter, the
sampler will send a reset MAD message after each counter query.

Change-Id: I65690af4d638cc29c8047d958fe5c1270c175902
NOTE: This is Narate's patch from the master