ldms: Add metric source 'status' into lustre2_* samplers.
Metric source 'status' is added into each of the lustre2_* samplers to help
indicate the availability of the source. status 0 means the source is
unavailable, and 1 means the source is available. For example,
if client.lstat.status#llite.flustre is 1, the client is mounted and the client
data can be obtained properly. Status being 0 means the sampler thinks the
metric source is unavailable. This might be because the lustre file system is
not mounted or simply because of sampler misconfiguration.

Change-Id: I3b99bd98dc4388ee8d386882ab2969bdf1fcadfa