ldms: Specify the to-be-stored metrics by their core name
If only a subset of metrics in a metric set should be stored, users need to give
the full metric names, <core>#<extension>. This is inconvenient. This change
allows users to give only the core part of metric names. That is, instead of
giving idle.rate#0,idle.rate#1 users can specify only idle.rate.  ldmsd will
store all metrics in the metric set with the core name of idle.rate.

Change-Id: I6fdb1644c175d905a7c7fd2e99f4c2903a00b3b6