ovis: Fix the use of pthread_cond_wait and pthread_cond_signal
pthread_cond_signal wakes up at least one thread so the condition to wait for
the signal should be a while loop instead of an if statement. This is because, if
at least two threads woke up, the condition might not be true when the second
thread acquires the lock.

Change-Id: Iced819af906f2ec35b81bf1b14ace0ea2a30fc2f