Fix B+ tree handling of duplicate keys
The B+ tree was not positioning iterators at the first key if there were duplicates.
This would make iteration over all matching keys miss keys that preceded it in
the index. This has been fixed. In addition, if the key is present, _lub will position
the iterator at the last matching key, and _glb will position the iterator at the
first matching key.

Signed-off-by: Tom Tucker <tom@ogc.us>

Change-Id: Idd5cdb122b82ff45185b06b6cb424a1192714a5e