komondor: non-blocking action execution with checking
komondor previously used 'system()' call to action script, which will block the
work of action handling thread. This change modify komondor to fork, exec and
handle action return code in SIGCHLD handling function instead. This is done so
that a long-running action won't prevent other actions in the queue.

In addition, the before the execution is called, komondor will check whether
there is an outstanding running action for the same event (same model_id,
metric_id, action). If there is an outstanding action, the new action will not
be executed.

Change-Id: I269ffee2847a37f0800db40e7ed2ebb238d808f7