lib: zap_sock - set keepalive parameters for initiator side socket.
In a zap application (ocmd), once a connection is established, the passive and
active sides talk back and forth a little initially. Then, they are both being
quiet. When the passive side went down gracelessly (e.g. in a hard shutdown
event), the active side doesn't receive any indication that the passive side is
down. When the passive side comes back up again, the active side doesn't
re-initiate the connection because it still think that the connection is OK and
passive side was there all the time.

This change address this problem by setting socket option SO_KEEPALIVE and their
related parameters (TCP_KEEPCNT, TCP_KEEPIDLE, and TCP_KEEPINTVL) so that a keep
alive probe will be send to the passive side when the connection is idle (10
seconds). It has been setup so that three probes are used, 2 seconds apart from
each other, to determine whether a connection is alive. So, in to total it would
take 16 seconds to determine that a link is down in this particular case.

Change-Id: I9c0ba35b8c1a6476e2839f40413c810dd6e1715b