sos: bpt - fix 'bad_ref' assertion bug.
In debuging mode, bpt's verify_tree() check for bad references (bad_ref) within
the tree. This is meant to be checked only in deletion case, but verify_tree()
is called from both insertion and deletion path. When ODS reuses the freed block
that happeneded to be an un-reset bad_ref, bpt raises an error. This changeset
add bad_ref reset after it is not needed.

Change-Id: I625cf0e495eeaf0154cf4344e46ef9460879178f