ldms: store_sos: Added new attributes when configuring store_sos
owner: Change all sos stores' owner after they are created.
metric_names: A comma-separated list of <metric_name(type)>, where 'metric_name'
is the store sos core name and 'type' is the type of the metric. If this
attribute value is given, the store sos of the given metrics will be created
when the store_sos plugin is configured.

Change-Id: I6dc64b4c07333170268ae7b50f06920f42e74258