ldms: Fix of ldmsd stuck when start over ssh/pdsh
When ldmsd open a log file, it did not redirect fd 1 and 2 (stdout and
stderr) to the newly-openend log file. In addition, when a user starts
ldmsd over ssh or pdsh, ldmsd stdout and stderr will be the same as
sshd. Even though ldmsd daemonized, stdout and stderr still point to
the those of sshd. This makes sshd session stall. This changeset use
dup2 to redirect stdout and stderr to the log to solve this problem.

Change-Id: If719ea16805f69e9613ce8ffc08481d1c76ffa78