Fix leaf confusion bugs in RBT
The earlier RBT used LEAF sentinels to initialize left/right of RBT nodes. This
allowed the logic to incorrectly rotate and otherwise modify LEAF nodes
corrupting the tree. Replaced the LEAF sentinel with NULL. This will cause
any incorrect treatment of a LEAF node to crash instead of silently corrupting
the sentinel.

Signed-off-by: Tom Tucker <tom@ogc.us>