ocm: oparser - add host fix
add host command generated by oparser had a problem of assuming the
'connection target' hostname (e.g. vic1-rdma) is the same as actual
hostname (vic1). This is wrong. In some systems, the name used for
initiating connection wasn't real hostname (the name that you can
obtained from the 'hostname' command). It could be just a convenient
name for different kind of connection (-eth, -ib, ...). This cause a
problem because the LDMS sampler side uses real host name to create a
metric set. Hence, the aggregator would fail to lookup the set
(because he was looking for a wrong set name).

This changeset fix this problem by introducing another tag named
'conn_target' to explicitly specify connection targets (like
vic1-rdma) in addition to real hostname. If 'conn_target' is not
specified, the hostname is used.

Change-Id: I579da2610ea68aa627dd708fa1fc0942e767a672