lib: build/doc permission fix
doc build rule in lib/Makefile.am was not invoked when make is called.
Instead, it was subsequently invoked when install-data-local rule was
invoked when 'make install' was called. This caused a problem in the
system that call 'make' and 'sudo make install' as 'build/doc' will be
generated at install time by root in the build directory.  The correct
behavior would be that 'build/doc' should be generated when 'make' is
called and 'make install' would just copy the documents to the
installation destination. This change set implements the
latter-described behavior.

Change-Id: I3c44b92138697fbb1920f88a34f6b528a3fd8a37