added fileperm check ala ssh for auth secrets.
Note that the permissions on a secret file are only
checked if the user is the owner. A user pointing at
a file they don't own is allowed to use that secret:
the non-user owner is assumed to know what they want for
security. This is how ssh does it.