ldms: zap replaces ldms_xprt
NOTE: rebased on top of ldms3

NOTE: This is expected to be a long-live change in Gerrit iterative
review process. So, this first one is an early design review. The
detail is as follows.

This change addresses the following:

- Replacing ldms_xprt with zap
- Modifying ldmsd, ldms_ls connect to zap non-blocking connect model

This has been tested on both sock and rdma. ldms_ls can query ldmsd
just fine.

Found bugs that will be addressed in another patch (or other changes)
- ldms_xprt.c:alloc_req_ctxt() leaking
- set descriptor (sd) leak
- "local" transport elimination

Change-Id: I4f2c663a6ee677de16a54456ed54413e85ff15f7