ldms: 2nd-level agg update_complete_cb race fix
update_complete_cb in the 2nd-level aggregator may be called with
meta_gn 0. This situation can occur when 1st-level aggregator finished
looking up the set in the sampler, but it is not done updating the
set. When this happenedh, 2nd-level aggregator will die due to mvec
assert in update_complete_cb() function.

This commit fix this bug by skipping update_complete_cb() routine if
meta_gn is 0.