ldmsd: Get rid of the duplicated code in ldmsd.c and ldmsd_ocm.c
The logic to control ldmsd, e.g., to load a plugin, to configure a plugin and to
add a host, is embeded in the functions that process the configuration received
from either ldmsctl (the process_* functions in ldmsd.c) or ocmd (the
ocm_handle_cfg_* in ldmsd_ocm.c). Thus, the code is duplicated in both files.

This changeset pulls out the duplicated code and makes them to be functions in
ldmsd.c. Now the process_* functions in ldmsd.c and the ocm_handle_cfg_*
functions in ldmsd_ocm.c only parse the configuration strings and then call the
new functions.

The ocmdsqlite3 plugin in ocmd is also modified to send the attribute values of
the type ldmsd expecting.

Change-Id: I48840e3632cb0e8391510fc1d4bf831c5addbf6d