ldms: ldms_lookup fetches meta-data
Change ldms_lookup so that it fetches the set meta-data

Previously, the set was not usable after ldms_lookup() except as
a parameter to ldms_update(). This was very confusing to some users.
This change makes it so that all meta-data related functions
can be used after ldms_lookup() completes.

Signed-off-by: Tom Tucker <tom@ogc.us>

Change-Id: I609fc30dc258780b6880fd2ffa12040dca6b5058