baler: bmem and store - add _refresh() interface.
For the bmem that is opened by multiple processes, the writer process
bmem will get remapped if necessary by bmem_alloc(). However, the
reader process does not remap the memory, and prone to accessing the
memory beyond the mapped region. *_refresh() interface is introduced
for the reader processes to check and remap if necessary.

Change-Id: I62d122dc52c62091b0c545f870552086dc25939d