ldms: bad configure generation fix
AC_CHECK_HEADERS and AC_CHECK_LIB seem to not liking the []
surrounding AC_MSG_FAILURE() macro. On Ubuntu (with autoconf 2.69), it
seems to work fine, but on CentOS 6.4 (autoconf 2.63) it doesn't work.
The problem was that if AC_MSG_FAILURE() was surrounded with [], the
resulting configure file is busted. This is likely to be an autoconf
bug. This changeset remove the surrounded [] so that the configure.ac
works with older autoconf.

Change-Id: I8bfebbd8aa20f64075d1b8b618d9ea2dc20650cf