ldmsd: Add the one-shot sample command
The 'oneshot' command is added. Upon receiving the command, ldmsd will schedule
a sample event on the sampler plugin at the specified time. The time can be
either a timestamp or a special keyword 'now+<second>.'

The oneshot command (sampler daemon) could be used in conjunction with the
notify command (aggregator) in order to store the updated sets by the oneshot
command.

Change-Id: I43c74733c3cdf6f7643b041132f1c2ef1b6afa37