sos: ods - extend_file() bug fix
extend_file() uses lseek() and write() to extend the file.  lseek()
return type is off_t, which is usually bigger than int as most target
systems are 64-bit.  Hence, the extension that makes the file a little
bit bigger than 2GB will make the integer return code variable
overflow. This will be mistakenly seen as a negative value and be
treated as seek error.

This commit fix the bug by using appropriate variable type for both
lseek() and write().

Change-Id: I83ed67917335749785a8d93c5bc8d8051d6341d1