ldmsd: Add log/verbosity level feature to ldmsd.
The -v command-line option is added to speicify the verbosity level.
The -q option for quiet is removed. ldmsd could be started with quiet mode by
giving -v QUIET at the command line.

The log message levels, in order of increasing importance, are DEBUG,
INFO, WARNING, ERROR, CRITICAL and SUPREME. The SUPREME level is a special level
that is for messages printed per requrests, .e.g, messages printed when ldmsd
receives the info command.

The ldmsd verbosity levels, in order of decreasing verbosity, are DEBUG, INFO,
WARNING, ERROR, CRITICAL and QUIET. For example, if users start an ldmsd process
with the DEBUG level, all messages will be reported. If users start the process
with the ERROR level, only messages with the levels of ERROR, WARNING and
SUPREME will be printed.

Change-Id: I0a1f8380707b7da6f32962f9133cdf5d26179d33