ldms: ldmsd: ldms_ls: Added ldms_xprt_with_auth_new() API and Auth. nit fixes
I made a bad decision when changing the ldms_xprt_new() API while adding the
authentication feature to ldms. In this changeset, I changes the ldms_xprt_new()
API back and adds a new API ldms_xprt_with_auth_new() for creating an ldms
transport with authentication. The API is available if ldms is built with the
--enable-auth option.

Change-Id: I2885f5090dc3afe03b985854ab456f18fa1c7e33