baler: img2 query - ts, comp_id wrong round off fix
In bhttpd img2 query, ts_end is calculated by
  ts_end = ts_begin + npp*width
which is mathematically correct. However, the expression is of float
type, and got round-off to int as ts_end is of type int. By losing the
precision, sometimes the result is wrong (e.g. ts_end < ts_begin).

This commit fixes the bug by casting the addition part to int:
  ts_end = ts_begin + (int)(npp*width)
The addition part should not exceed the precision of the mantissa.
Hence, we yield the correct result.

The host_end calculation also follow the same logic above.

Change-Id: I196027e39f9c7d98dd4e09879fd37fffabd4538d