ldmsd: Add dynamic configuration objects
This change adds dynamic configuration to LDMSD. The original
configuration will be left intact until the full array of
replacement function is implemented and reviewed.

This patch adds generic Configuration Objects. A Configuration Object
is a named, reference counted and locked base object that the
type-specific sub-types are based on.  There are three sub-types:
Producer, Updater, and Storage Policy.

A Producer is an instance of an LDMSD.  It is used to manage the
connection between two LDMSD: an Aggregator and a Sampler. A new
connection management subsystem is added for this purpose
in the file ldmsd_prvdr.c. Connection management and sampling are
decoupled in order to simplify the state machines and improve error
recovery for these two activities.

An Updater is a remote sampling policy. It consists of a collection
of Producers and parameters that specify how frequently to sample
and which sets.

A Storage Policy specifies which sets are stored. It specifies which
Producers, Sets, and Metrics are passed to Storage Back Ends when
an update completes.

A new python script has been added to submit the new configuration
commands. This command supports command complete, help and other
features to make it easier to use the new configuration subsystem.

Fixed a bug in the udata that would not be seen by the peer when
the value was changed because the meta_gn was not being updated
when the value was changed and do_read_all() would not be called.

Change-Id: I56aba0197a0570a737e16a113ef83b3d7beb883f