baler: GUI - HeatMapDisp movement freeze fix
HeatMapNavCtrl can set any timestamp and component ID for the top-left
pixel of the image. If the given values are out of range, HeatMapDisp
will still show the image according to the navigation command.
However, when the HeatMapDisp is dragged, the image location will not
change. This is due to the logic in HeatMapDisp::sanity() function to
not move when the top-left corner is not in a correct range.

This commit fixes the problem by allowing the out-of-range movement if
it is moving towards the range.

Change-Id: I88239a99b3d4c57907fee846f27f95cce65da25a