ldmsd: Prevent creating the 2nd set if the sampler does not support.
Added the logic to check if a set is already created before creating a new
one in the sampler plugins that do not support multiple sets. Specifically the
sampler plugins that define the set handle as a global variable. An error is
returned if the set handle is not NULL.

Change-Id: I0d63777b623d1921ad508b598885f24c3e24960f