ldmsd: ldmsd_controller: Added --source and --script cmd-line options
The command-line option --source is added so that users can give the absolute
path to a config file wihout starting the interactive interface.

The command-line option --script is added so that users can give the absolute
path to a bash script that generates ldmsd commands and the script arguments.
Similar to the source command, users can give the script path and its arguments
to the command script in the interactive interface. The script command will
executes the given script and its arguments and then forward the script output
to the ldmsd.

Change-Id: I710f78a18a8efaf78476dca7ca13b49160498478