ldmsd, ldmsd python: info command improvement
An optional attribute, called 'name', is added to the info command. The value
could be one of the prdcr, updtr or strgp. If the name attribute is given, only
the specified information will be reported to the log file.

Example:
     info name=prdcr
will report the states of all producers to the log file.

Change-Id: I011bccb0c7a133af6efa4eda40b126e3e3678ad4