Add support for kernel LDMS (KLDMS)
Support includes schema, set, and metric management that is exported so
that kernel samplers can be provided. The kldms modules also provides a
character device interface so that application can map access to the sets.
There is also a /proc interface that can be used for management.

Finally a vmstat kernel sampler is provided.

The code can be built a such:
	# make -C path_to_kernel_source SUBDIRS=path_to_kldms_source

Change-Id: Idc206d641a24d14020f709cf78183d9df226dff9