Add KLDMS sampler for GPCD
There is a new kernel module kldms_gpcdr that creates a LDMS schema and
metric set. The metrics are populated by calling a read register function
that is in Cray's GHAL. The HAL functions have been stubbed out in ghal.c
and ghal.h. Eventually both of these files will be removed.

gpcdr.c is paired down code that was received from Cray. It reads the
registers into arrays containing the metric. There is sampling throttling.

kgpcdr is the KLDMS sampler. It also creates the schema and set.

Change-Id: I556dc0146b08e0deb37394793b694213bc703f13