baler: slave token attribute bug fix.
Fix slave token attribute bug. The symptoms of the bug is that the
slave balerd pattern extractor does not mark the '*' field correctly.
The reason is that when the slave make a request to insert-or-query to
the master balerd, it sets token attribute to BTKN_TYPE_LAST,
expecting the master to fill in the attribute. The master, however,
wrongly check the attribute assign condition. The master would set the
attribute as slave requested if the attribute was not BTKN_TYPE_STAR.

This patch fix the condition on the server side to check for a correct
condition.

Change-Id: I072e8940ba58e10c68a09da98af1f9beee35bec1