Fix the metric order in ldmsd_strgp to match the set
The metrics were being added to the head of the list which results
in the metric list on the storage policy being in the reverse order
of the metrics in the set. Also, the metric list was not being passed
to the open call in the store, so it was not possible to know what
metrics would be in a subsequent store call.

Signed-off-by: Tom Tucker <tom@ogc.us>