Fixed a few ldmsd configuration problems
The ldmsd_config.py script had a base class definition that had socket
as a function, but a subclass that had it as a member variable. If
an exception is thrown in __init__ sockpath issues, for example, the
error doesn't represent the real problem.

The code assumed the sockpath included a '.' character.

The ldmsd_config.c info had too many columns for the strgp output.

Updated the prdcr warning to recommend using the -m option if a
set could not be created due to memory resource exhaustion

Signed-off-by: Tom Tucker <tom@ogc.us>