fixed procstatutil2 and flatfile problems found on chama.
Flatfile is now usable in v2, where before it was crash prone.
Also fixed several places where return codes were incompletely handled and
not logged. Fixed one missing lock that leads to incorrect flush count badness.
Added extra gap for readability in synthetic data output.
Lowered the usual condition of a node not yet reporting from ERROR to INFO
so as not to drown syslog with messages unless we are in a diagnostic mode.