Max the config string length configurable
The configuration string length was hard-coded to 8192 on the ldmsctl utility
and 256B in the ldmsd. Ironically, there is a #define in ldmsd.h called
LDMSD_MAX_CONFIG_STR_LEN, however, it was ignored by the code.

Bumped the default size of this #define to 8192B, used the #define when
allocating config string buffers, and added logic to test for the
LDMSD_MAX_CONFIG_STR_LEN environment variable and use it to
override the #define if present.

Signed-off-by: Tom Tucker <tom@ogc.us>