ldms: ldms_dir last set stall fix
ldms_dir_reply() traverse through RBT, and put set names into the
send buffer. When a set cannot be put into the buffer, the buffer is
flush, and the set is put into the clean buffer, waiting to be sent in
the future node visit. However, if such set is of last node in the
tree, the buffer is never flushed, causing ldms_ls to be stalled as it
is expecting ldmsd to send more set names. This commit fix this
problem.