rpm7 WITH option fix
This commit fixes the bug that fails rpm check-buildroot.

Rpm check-buildroot greps for <RPM_BUILD_ROOT> on all files under
<RPM_BUILD_ROOT> (except *.pyo, *.pyc *.elc and .packlist), and fails
if there is a file containing the string that matches
<RPM_BUILD_ROOT>.

Because rpm7.sh blindly applied all '--with-XXX' flags to all
sub-projects, the --with-ovis-lib=<OVIS_LIB_BUILD_ROOT> is
supplied to configure of ovis/lib as well, resulting in having
<OVIS_LIB_BUILD_ROOT> (RPM_BUILD_ROOT) appearing in lib-pedigree which
fails rpm check-buildroot.

This commit solve the issue by making rpm7.sh correctly supplies
appropriate WITH arguments to the configure script in each
sub-project.