Remove the producer name mismatch log warning
The prdcr_lookup_cb() function was printing a warning message if
the producer name in the set did not match the rule name. This
warning is not helpful because the rule name doesn't have any
logical meaning with respect to the set's producer name. For
aggregators, this warning was almost always printed because the rule
name could only ever match one of the computes from which it
was aggregating.

Signed-off-by: Tom Tucker <tom@ogc.us>