lib: zap - completing event interposing
CONN_REQUEST event was the only zap event that is left not interposed,
due to sock and ugni drivers that expect the application to choose to
accept or reject the endpoint before the CONN_REQUEST application
callback is returned.

This commit complete the event interposing by changing the drivers to
not expect accept/reject right after CONN_REQUEST event notifcation.