baler: rsyslog parser negative timezone minute bug fix
The current timestamp calculation in rsyslog parser is wrong for the
negative timezones with non-zero minutes (e.g. Venezuela), as the
calculation did apply the negative sign to the minute part. For
example, -4:30 is treated as -4:00 +0:30, which should have been -4:00
- 0:30.

This commit fix the bug by correctly apply the negative sign to the
minute part of the time zone.