balerd: slave message processing credit
For each input message that the slave process, balerd will allocate
memory to track processing state of the message. The slave send
requests to the master to resolve unknown token or pattern. The memory
will be freed in the last step of the slave message processing after
the slave resolved all unknown tokens/pattern for the message.

It is possible that the slave's request rate is greater than master's
resolve rate. In that case, the slave's memory will keep growing
without bound. The slave message processing credit is used to limit
the number of active processing messages, keeping the context memory
from growing without bound.