zap_ugni: Do not deliver the disconnected event until successfully unbind
The mm_info memory has been stomped, and the cause could be that the ugni copied
data to the mapped memory after zap_ugni delivered the disconencted event to
ldms. This would corrupt the the mm_info memory.

We witnessed this when the GNI_EpUnbind returned with the gni error NOT_DONE.

With the change, zap_ugni will not deliver the disconnected event until it can
successfully unbind the gni endpoint or it has been trying for a user-specified
duration.

Users could set the ZAP_UGNI_UNBIND_TIMEOUT environment variable to specify how
often zap_ugni tries to unbind an gni endpoint. ZAP_UGNI_DISCONNECT_EV_TIMEOUT
is the environment variable to set how long zap_ugni will try to unbind the gni
endpoint until it gives up, delivers the disconnected event to the application,
and cleans up the zap endpoint.