ldms: ldms never remove a set from the tree and destroy
An aggregator could read garbage from its producer if the set on the producer has been
deleted while the aggregator is receiving the dir del message.

With this commit, a set will always exist after it is created. ldms_set_delete
only unpublish the set when the remote rbd list is empty.

Change-Id: I932e55a62317009cb7ce49519f513c6cc9c5cfb0