baler: balerd - multiple output queues
Previous attempt to parallelize output entry processing is still
busted (thanks to Tom who pointing this out) as the slowest plugin
took a long lock while fast plugins quickly process the other entries,
until the front of the queue is filled with output entries for the
slowest plugins. This will block other output worker threads as the
output plugin took a lock while it is processing the data.

This commit give each output thread its own output queue. An output
plugin is assigned to the least-busy output queue at the time it is
created. In this approach, an output worker thread won't be blocked
by the other output worker thread.

NOTE: The slowest output plugin still holds everybody off as it blocks
the input processor (who create output entries) when the slowest
output queue is full. But, we will have real multi-threaded working at
the same time when the queue is not full. The performance is better
than that of the previous approach (~ 16% improvement).