ldmsd,ldms_ls: Get rid of the -a none option
To prevent any confusions how to disable authentication in ldmsd, the -a none
option is removed. With the change, enable/disable authentication is a compile
time configuration.

The default is to compile with authentication. To disable authentication in
ldmsd, users will need compile ldmsd with --disable-ovis-auth.

The ldmsd greeting message includes the infomation whether ldmsd is started with
or without the authentication.