removed the syntactic sugar on flt/dbl endian swap
The previous casting just makes gcc whinge about type punning.
There are two underlying worries: the union layout and the ieee (or not)
byte order when endian reordering on non-x86/gcc platforms.
We use the integer reorder operators on floating point and
provide a separate test_metric (other commit) to verify it.

Because the code for float/double endian conversion may on some
platform not be correct, it should not be merged with known
to be correct code in other branches of the switch statement.