ldms: ldmsd exit code 0
When ldmsd received a termination signal (SIGHUP, SIGINT or SIGTERM),
it calls `cleanup(100)`, which finally calls `exit(100)`. This
confuses systemd to think that ldmsd exit with an error, hence setting
the status of ldmsd service as error.

This commit changes the cleanup call code from 100 to 0 as this is a
normal daemon termination.