Fix schema name length calculation and first metric address logic
The logic that computed the schema_name and first_metric_desc
addresses was adding the size of the ldms_name_t structure twice.
The meta->meta_sz value, however, did not account for these extra
bytes.

Also, the instance_name length included the terminating '\0', but
the schema_name length did not.