Add check for NULL value in storekeys table
The change to cleanup memory at rollover introduced a bug
that results in a crash in close_store when there is more
than one open CSV store handle.

The code that was added frees the matching storekey entry
and sets the entry to NULL. A subsequent call to close
will crash when it attempts to compare the NULL storekey
entry to the key in the store handle.