PAPI - Change the application name file structure
The previous structure was the application name, job, and
user name placed in three lines in the file. while the new
structure now is having one line in the file that contains
the three parameters separated by a comma.
This change will eliminate the error where the job is changed
but the file content is not