The inet socket configuration handling is single-threaded
There is a single thread that handles both the connect/accept logic
and the processing of configuration requests. This prevents multiple
clients from interacting with the ldmsd. In practice, this
prevents the GUI from getting ldmsd status when a user is in the
ldmsd_controller and conversely precludes anyone from running the
ldmsd_controller when the GUI is running.

The failure mode also fills the log file and eventually the filesystem
with error messages because the accept path on error is a goto/retry loop.
This change limits the enumber of error messages to 10 between successful
connects and adds a sleep after an error to prevent the error path
from consuming 100% of the CPU.