zap_rdma - endpoint resource leak fix
The cause of the resource leak was that cq_channel was only enabled
(add to event poll) after the connection is established. Hence, when
the connection was not successful, the recv work entry posted at
connection setup time won't be delivered. Since each recv work entry
took a zap endpoint reference, the endpoint will never be destroyed.

Thus, this patch enables the cq_channel in the event of rejection and
connection error. This has been tested with zap_test,
zap_test_reconnect, and ldmsd aggregator (with and without sampler
daemon).