ldmsd: update auto-interval
An updater has a default schedule specified thru configuration. If
a producer set that is updated by the updater has an update hint
different from the default schedule, the updater will update such
set according to the hint contained in the set.

The hints are either the sampling interval and offset if ldmsd is
a 1st-level aggregator or update interval and offset if it is an
aggregator of a higher level.

The interval and offset hints are assumed to be in microseconds.

If a hint contains an offset value, the update offset will be
'offset hint' plus 100000 by default. The increment of the offset
can be altered by exporting the environment variable
LDMSD_UPDTR_OFFSET_INCR. The unit is in microseconds.