ldms: ldmsd_task bug fix
ldmsd_task state wasn't properly changed to STARTED. This results in
incorrectly initializing ovis_event that has been added to the
scheduler, and re-adding it into the scheduler. When the event is
removed, ovis_scheduler will experience segmentation fault due to bad
memory access.

This bug is found by Nichamon.