ldms: set publishing
- export `ldms_set_publish()` and `ldms_set_unpublish()` APIs.
- `ldms_set_new()` doesn't automatically publish the set. The
  application needs to call `ldms_set_publish()` to publish the set.

NOTE: All `ldms_set_new()` calls in ldmsd samplers are affected by
this behavior change. However, this patch only addresses the call in
`sampler_base` because sampler plugins should be based on it.  We're
in the process of maintaining/deprecating samplers. The supported
samplers will be refactor to be based on `sampler_base`.  Otherwise,
it will be removed.