ldmsd: fix the bug that updater offset cannot be disabled at udptr_start line
If an offset value is given at updtr_add line, the updater will update sets
synchronously although the offset isn't given at the updtr_start line. The patch
changes this behavior so that, if an interval value is given at the updtr_start
line but not the offset value, ldmsd will interpret that users want the updater
to update the sets asynchronously.