kokkos build fix
According to
https://www.gnu.org/software/automake/manual/html_node/Yacc-and-Lex.html
the generated C files and header files from lex and yacc are included
into the source distribution (`make dist`). As a result, `make dist`
failed if kokkos is not enabled because either automake LEX/YACC
support is not initialized, or the system does not have LEX/YACC
installed.

The solution in this patch is to:
- explicitly list `.y` and `.l` files in `EXTRA_DIST`,
- explicitly add rules to generate C sources/headers from Lex/Yacc
  files,
- add the generated C files in `nodist_` list so that they won't be
  included in the distribution.