ldmsd - process_config_file error exit bug fix
config request handlers doesn't always return non-zero code if the
processing failed. Historically, I believe that the handler return
code was meant to be for something else (like resources / condition in
the processing side), and the response error code was meant to be the
one that really report the error in the configuration input. As a
result, some configuration error won't make ldmsd to exit due to rc 0,
with error response code.

This patch uses the response code and return code to make the
configuration process stops and ldmsd exits.