Fix ldms strgp_prdcr_add regex match
A strgp can be associated to a producer set in two paths: lookup
callback path (strgp is created before lookup is completed), or strgp
start path (strgp is created after a lookup has been completed). The
former path associated the strgp with a prdcr set without checking the
regular expression, which is addressed in this patch.