Check for processed sets in `ldms_ls`
With ldmsd_setgroup, `ldms_ls` could report (with -v or -l) a set
in a group multiple times if the user does not specify any set list,
or provide a set list that contain both group and the member (e.g.
`ldms_ls mygrp set0 set1`). To prevent this, this patch introduce a
red-black tree to hold the names of the set that has been processed by
ldms_ls.