lustre sampler - stat sampling logic fix
Lustre sampler stat sampling logic was comparing the metric unit with
`regs` (which was a typological error, the correct token is `reqs`) to
determine whether the `count` or `sum` should be used. According to
http://wiki.lustre.org/Lustre_Monitoring_and_Statistics_Guide#Stats
a stat line is in one of the followig formats:
- {name} {count of events} samples [{units}]
- {name} {count of events} samples [{units}] {min} {max} {sum}
- {name} {count of events} samples [{units}] {min} {max} {sum} {sum2}
(where sum2 is sum-of-squared).

This patch changes the sampling logic to use `sum` when available, and
use `count` when `sum` is not available to collect counter data for each
metric (e.b. byte counter).