Fix memory leak in the sock transport
The send queue is not scrubbed when the endpoint
is destroyed. This leaves incomplete writes on
the queue and leaks memory on failed connect
attempts and other errors.

In z_sock.h, the pack pragma is set to 4 which
results in  2B of space between msg_type and
msg_len sock_msg_hdr. This space is unitialized
and contains garbage. Setting the pack properly
would also fix this, but would break
compatability.