ldmsd_config error handling for deferred starts
Errors in deferred start commands were ignored by the config file
processing procedure. Hence, when an error occurred in the deferred
start commands, the configuration process continued without any warning
in the log file. This patch enables various warnings as well as stop the
config file processing when encountering the error in the deferred start
commands.