zap_sock: zap event queue - closing xprt deadlock fix
When the zap event queue is filled up, `zap_interpose_cb()` (which is
`sep->ep.cb`) cond_wait for a vacancy. When an application calls zap
operation (e.g. zap_unmap()) from the proper zap event callback thread
(the consumer of zap event queue), the operation also lock the same
endpoint mutex being long-time locked by `sock_event()` while flushing
socket ioq (that is blocked -- waiting for a vacancy). This resulted
in a deadlock.

This patch unlock the endpoint mutex before (blocked) posting the zap
event to the event queue so that the consumer thread can finish the
callback and process the next entry -- making vacancies in the queue.

NOTE:
- `zap_interpose_cb()` (or ep.cb()) does not need `ep.lock` to be
  held. In fact, for zap_sock, the `ep.cb()` inside this patch is the
  only place being called with `ep.lock` held.
- `zap_interpose_cb()` also (correctly) took endpoint reference.
- the code path inside this patch already has an endpoint reference
  held. So, no need to take extra endpoint reference before calling
  `ep.cb()`.